+++ pypalex/__init__.py
@@ -1,26 +1,29 @@
-"""!
-#######################################################################
-@author Al Timofeyev
-@date   March 3, 2022
-@brief  Initialization file for PyPalEx.
-#######################################################################
-"""
+##  @file   __main__.py
+#   @brief  Initialization file for PyPalEx.
+#
+#   @section authors Author(s)
+#   - Created by Al Timofeyev on March 3, 2022.
+#   - Created by Al Timofeyev on April 7, 2023.
 
 from .settings import __version__, __cache_version__
 from . import Extractor
 from . import arg_messages
-from . import extraction_utils
-from . import conversion_utils
-from . import image_utils
 from . import constants
+from . import conversion_utils
+from . import extraction_utils
+from . import file_utils
+from . import image_utils
+from . import print_utils
 
 __all__ = [
     "__version__",
     "__cache_version__",
     "Extractor",
     "arg_messages",
-    "extraction_utils",
-    "conversion_utils",
-    "image_utils",
     "constants",
+    "conversion_utils",
+    "extraction_utils",
+    "file_utils",
+    "image_utils",
+    "print_utils",
 ]