+++ pypalex/settings.py
@@ -3,21 +3,22 @@
 #   @details    Used to set up the version numbers and
 #               default output locations.
 #
+#   @note   This code has been borrowed from Dylan Araps
+#           PyWal on github: https://github.com/dylanaraps/pywal/blob/master/pywal/settings.py
+#
 #   @section authors Author(s)
 #   - Created by Al Timofeyev on March 2, 2022
 #   - Modified by Al Timofeyev on April 21, 2022.
 #   - Modified by Al Timofeyev on March 11, 2023.
 #   - Modified by Al Timofeyev on March 22, 2023.
 #   - Modified by Al Timofeyev on March 26, 2023.
-#
-#   @note   This code has been borrowed from Dylan Araps
-#           PyWal on github: https://github.com/dylanaraps/pywal/blob/master/pywal/settings.py
+#   - Modified by Al Timofeyev on April 7, 2023.
 
 
 import os
 import platform
 
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 __cache_version__ = "1.0.0"
 
 HOME = os.getenv("HOME", os.getenv("USERPROFILE"))